requirements.txt
streamlit==1.28.1
openai==1.35.0
langchain==0.2.11
langgraph==0.1.9
pandas==2.1.3
numpy==1.26.4
plotly==5.17.0
python-dotenv==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
seaborn==0.13.0
matplotlib==3.8.2
scikit-learn==1.3.2
altair==5.1.2